Add files via upload
In this checkpoint, I created a presentation where I explained the following subjects:
How does the web work?
What do I need to be a web developer?
Why did I choose to learn web development?
A day in the life of a web developer?

The aim of this is to learn how to research, ask precise questions, and find the best answers. In other words, the goal here is to learn how to learn and prepare my mind towards becoming a web developer in the near future.